# Conflicts:
#	Engine/.vs/Engine/v15/.suo
#	Engine/Engine/Code/Framework.cpp
#	Engine/Engine/Code/Framework.h
#	Engine/Engine/Code/Renderer/Objects/Mesh.h
#	Engine/Engine/Code/Renderer/ShaderPrograms/PostProcessingShader.cpp
#	Engine/Engine/Code/Renderer/ShaderPrograms/ReflectionMapShader.cpp
#	Engine/Engine/Code/Utils/ModelLoader.h
#	Engine/Engine/Engine.ilk
#	Engine/Engine/Engine.pdb
#	Engine/Engine/Engine.vcxproj
#	Engine/Engine/Engine.vcxproj.filters
#	Engine/Engine/Particles/dust.json
#	Engine/Engine/Shaders/pixelDeferredLightning.ps
#	Engine/Engine/Shaders/pixelForwardPlanar.ps
#	Engine/Engine/imgui.ini
#	Engine/Engine/x64/Debug/Engine.log
#	Engine/Engine/x64/Debug/Engine.tlog/CL.command.1.tlog
#	Engine/Engine/x64/Debug/Engine.tlog/CL.read.1.tlog
#	Engine/Engine/x64/Debug/Engine.tlog/CL.write.1.tlog
#	Engine/Engine/x64/Debug/Engine.tlog/Engine.write.1u.tlog
#	Engine/Engine/x64/Debug/Engine.tlog/link.command.1.tlog
#	Engine/Engine/x64/Debug/Engine.tlog/link.read.1.tlog
#	Engine/Engine/x64/Debug/Engine.tlog/link.write.1.tlog
#	Engine/Engine/x64/Debug/vc141.idb
#	Engine/Engine/x64/Debug/vc141.pdb

+++ Engine/Engine/Code/ComponentSystem/World.cpp
@@ -20,8 +20,7 @@ void World::Update()
 	// update all components one type at the time
 	for (int i = 0; i < NUM_COMPONENT_TYPES; i++) 
 	{
-		unsigned int size = (int)_components[i].size();
-
+		size_t size = _components[i].size();
 		for (int y = 0; y < size; y++)
 			if (_components[i][y]->IsActive())
 				_components[i][y]->Update(delta);
@@ -50,8 +49,7 @@ void World::DeleteEntity(Entity* entity)
 // check if we have any components flagged for removal
 void World::HandleDirtyComponents() 
 {
-	int count = (int)_componentsToRemove.size();
-
+	size_t count = _componentsToRemove.size();
 	if (count == 0)
 		return;
 
@@ -66,8 +64,7 @@ void World::HandleDirtyComponents()
 // check if we have any entities flagged for removal
 void World::HandleDirtyEntities()
 {
-	unsigned int count = (int)_entitiesToRemove.size();
-
+	size_t count = _entitiesToRemove.size();
 	if (count == 0)
 		return;
 
@@ -85,8 +82,7 @@ void World::Empty()
 	// delete all components and entitys in the world
 	for (int i = 0; i < NUM_COMPONENT_TYPES; i++) 
 	{
-		unsigned int size = (int)_components[i].size();
-
+		size_t size = _components[i].size();
 		for (int y = 0; y < size; y++)
 			delete _components[i][y];
 
@@ -94,8 +90,7 @@ void World::Empty()
 		_components[i].clear();
 	}
 
-	unsigned int size = (int)_entities.size();
-
+	size_t size = _entities.size();
 	for (int i = 0; i < size; i++)
 		delete _entities[i];
